Feature/client role permission services

+++ internal/dto/permission/create.go
@@ -7,7 +7,7 @@ type CreatePermissionReq struct {
 }
 
 type CreatePermissionRes struct {
-	Id          int64  `json:"id"`
+	Id          int32  `json:"id"`
 	Name        string `json:"name"`
 	Code        string `json:"code"`
 	Description string `json:"description"`
